admin_app/requirements.txt
firebase-admin==6.5.0
fpdf2==2.7.9
qrcode[pil]==7.4.2
Pillow==10.3.0
python-dateutil==2.9.0
pyinstaller==6.8.0
cryptography==42.0.7
requests==2.32.3
